HADOOP-19164. Hadoop CLI MiniCluster is broken
